pyproject.toml
[build-system]
requires = ["hatchling", "cython>=3.0", "setuptools"]
build-backend = "hatchling.build"

[project]
name = "pop-pay"
version = "0.8.9"
description = "The runtime security layer for AI agent commerce. Drop-in CLI + MCP server — blocks hallucinated purchases and keeps card credentials out of agent context. it only takes 0.1% of Hallucination to drain 100% of your wallet."
authors = [{name="Point One Percent Team"}]
dynamic = ["readme"]
license = {text="MIT"}
requires-python = ">=3.10"
keywords = ["cli", "command-line", "agent-tool", "payment-cli", "browser-agent", "mcp", "mcp-server", "model-context-protocol", "ai-agent", "payment", "security", "guardrail", "virtual-card", "cdp", "browser-automation", "checkout", "prompt-injection"]
dependencies = [
    "cryptography>=46.0.5",
    "pydantic",
    "python-dotenv",
]

[project.urls]
Homepage = "https://github.com/100xPercent/pop-pay-python"
Repository = "https://github.com/100xPercent/pop-pay-python"
Issues = "https://github.com/100xPercent/pop-pay-python/issues"

[tool.mcp]
mcpName = "io.github.100xPercent/pop-pay"

[project.optional-dependencies]
llm = ["openai", "tenacity>=9.1.4"]
stripe = ["stripe"]
langchain = ["langchain-core"]
mcp = ["mcp>=1.26.0"]
dashboard = []
dev = ["pytest", "pytest-asyncio"]
browser = ["playwright>=1.40.0"]
vault = ["cryptography"]
passphrase = ["keyring"]
all = ["pop-pay[llm,stripe,langchain,mcp,dashboard,browser,vault,passphrase]"]

[project.scripts]
pop-launch = "pop_pay.cli:main"
pop-init-vault = "pop_pay.cli_vault:cmd_init_vault"
pop-unlock = "pop_pay.cli_unlock:cmd_unlock"
pop-pay = "pop_pay.cli_main:main"
pop-pay-doctor = "pop_pay.cli_doctor:main"

[tool.setuptools.dynamic]
readme = {file = "README.md"}

[tool.hatch.metadata.hooks.custom]
path = "hatch_build.py"

[tool.hatch.build.targets.wheel]
packages = ["pop_pay", "dashboard"]
artifacts = ["pop_pay/**/*.so", "pop_pay/**/*.pyd"]

[tool.hatch.build.hooks.custom]
path = "build_ext.py"

# [tool.cython] — Cython build process notes:
# - pop_pay/engine/_vault_core.pyx is compiled via setup_cython.py (invoked by build_ext.py)
# - The POP_VAULT_COMPILED_SALT env var (GitHub Actions secret) is injected into _vault_core.pyx
#   before compilation, then restored. This keeps the secret out of the repo.
# - Source/OSS builds compile with _COMPILED_SALT = None and fall back to the public salt.
# - cibuildwheel (build-wheels.yml) builds platform wheels for Linux, macOS, and Windows.

[dependency-groups]
dev = [
    "cython>=3.2.4",
    "setuptools>=82.0.1",
    "vulture>=2.15",
]
